+++ src/source/raster_dem_tile_source.ts
@@ -12,7 +12,7 @@ import type {Source} from './source';
 import type Dispatcher from '../util/dispatcher';
 import type Tile from './tile';
 import type {Callback} from '../types/callback';
-import type {RasterDEMSourceSpecification} from '../style-spec/types';
+import type {RasterDEMSourceSpecification} from '../style-spec/types.g';
 
 class RasterDEMTileSource extends RasterTileSource implements Source {
     encoding: 'mapbox' | 'terrarium';