+++ src/source/source.ts
@@ -41,7 +41,6 @@ export interface Source {
     attribution?: string;
     roundZoom?: boolean;
     isTileClipped?: boolean;
-    maplibreLogo?: boolean;
     tileID?: CanonicalTileID;
     reparseOverscaled?: boolean;
     vectorLayerIds?: Array<string>;
@@ -84,7 +83,7 @@ import video from '../source/video_source';
 import image from '../source/image_source';
 import canvas from '../source/canvas_source';
 
-import type {SourceSpecification} from '../style-spec/types';
+import type {SourceSpecification} from '../style-spec/types.g';
 
 const sourceTypes = {
     vector,